Add files via upload
Radar in a broader perspective entailed an electronic system that uses electromagnetic signals to detect positions of either dynamic or static objects. However, here, the radar system with the help of an ultrasonic sensor utilizes an ultrasonic signal in place of electromagnetic signals. In this work, an ultrasonic radar system that can detect and measure the position of an object in a short-range is implemented using Arduino Uno, ultrasonic sensor, servo motor, gun, and software. The object recognition range of the ultrasonic radar is within a distance of 40cm and the robotic angular rotation of the ultrasonic radar is from 2 to 178 and 178 to 2 degrees. Arduino Uno controls the radar system with a written code in the Arduino IDE environment and serially communicates with the display panel (computer) through the processing software. And after detection is complete the Arduino with instruct the gun to state shooting the detected object.